It works!
Compiles, runs as a service, starts and stops miners and starts IdleMon in the user's session.
Fixed VerifySettings and VerfiyGpu/CpuMiners.
Added some Debug output. Now outputs Log and Debug to Console session when not running as SYSTEM.
Added a lot of Debug logging.
Added miner executable checking, as well as checking for IdleMon.exe. In the future, may bundle IdleMon.exe into the app as a Resource item.
Added  MinerShouldBeRunning. Not certain if this method is going to remain.
Fixed checking for running miners, as well as starting them if not running and hopefully switching them to IDLE mining mode, though not tested.
Fixed correctly detecting a battery (hopefully), will test with a laptop soon.
Still need to test CPU Usage.
Still need to implement Pipe commands to pause/stop mining.