Temperature Monitoring!
Changed default json to MiningService.json. It should copy your old settings to the new file and delete the old file.
Added an Update button to the MinerEditor for saving current selected miner's settings.
Added temperature monitoring using LibreHardwareMonitor (aka OpenHardwareMonitor fork), as well as new options for cpu/gpu being throttled when themps go over the limits.
Added a setting for specifying a percent to wait for temperatures to drop before starting to mine again.
Added HardwareMonitor.cs for temperature monitoring.
Added config option to resume mining when locking or logging off of computer.
Updated the logic for checking if network connectivity exists, as well as subsequently moving the service startup logic to a separate task so we can return faster and the service won't fail to start.